From: info@sprow.co.uk
WIN32llseek sign extension bug

in the original code the off_t cast of 'lo' propagates bit 31 of 'lo' in
the widening cast to loff_t (also a signed number). So a lot of the time
this will go unnoticed (where the image size doesn't set bit 31), sizes
around 2G however fail as the result is overall negative.